Add script to authenticate against MediaWiki sites
MediaWiki protects against Login CSRF using a login token generated
on viewing the login page and storing it in the session and in a
hidden field in the login form. On submitting the login form, the
submitted token and the one in the session are compared to prevent
login CSRF. So, we need to first get the login token and session
cookie, then use those to perform the login request.