Code Review Updates
Updated alertRisk, alertTitle and alertConfidence to an array to give
flexibility in displaying internal vs external IP address

Updated IPv6 local address check to use RFC4193

Added isExternal function. This call is made only if the isLocal()
function call returns false.  The idea here is that it might not be a
local IPV6 or IPv4, but it still could be a valid external IP.  This
will test for that.  If it returns true, we alert the a slightly
different message.  If false, we likely have a formatting problem or
the F5 cookie is encrypted.